requirements.txt
Django==2.1
psycopg2
djangorestframework==3.7.7
flake8==3.5.0
exponent_server_sdk
rollbar
self
requests